# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Carga Horaria - API",
        "description": "Documentación completa de la API del sistema de gestión de carga horaria docente",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/rooms": {
            "get": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU10 - Listar aulas",
                "operationId": "bea681612754126f01807ff95647eb34",
                "responses": {
                    "200": {
                        "description": "Lista de aulas"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU10 - Crear aula",
                "operationId": "cd6c1029f22e2425a1f99cf83e27383a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aula 101"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 40,
                                        "nullable": true
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Edificio A, Piso 1",
                                        "nullable": true
                                    },
                                    "resources": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aula creada"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU10 - Ver aula",
                "operationId": "91a59d880bb0297e8321dd75af8f506b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU10 - Eliminar aula",
                "operationId": "61024891be80949c0aaa6e4536381225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula eliminada"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU10 - Actualizar aula",
                "operationId": "a0f53b42dd7a07e588b619d3c2031d6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "location": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "resources": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula actualizada"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/rooms/{id}/equipment": {
            "get": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU11 - Ver equipamiento del aula",
                "operationId": "6d44552c34f48a64780faf22174aeaa1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipamiento del aula",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "equipment": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Aulas"
                ],
                "summary": "CU11 - Actualizar equipamiento del aula",
                "operationId": "f76d388fd6b70ea612426e0fce431830",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "equipment"
                                ],
                                "properties": {
                                    "equipment": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Proyector",
                                            "Pizarra digital",
                                            "30 computadoras"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipamiento actualizado"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}